+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Deleting an event now correctly removes subject_event associations
 - Fixed a bug that incorrectly formatted column names in exports for domains
   with negative values
+- Fixed display of designs and sites on project reports
 
 ### Refactoring
 - Removed deprecated_grid_variables from variables